+++ XFaction/Configuration/Support.lua
@@ -45,7 +45,7 @@ XFG.Options.args.Support = {
 	args = {
 		Bar = {
 			order = 1,
-			name = format("|cffffffff%s|r", XFG.Version),
+			name = format("|cffffffff%s|r", type(XFG.Version) == 'string' and XFG.Version or XFG.Version:GetKey()),
 			type = 'header'
 		},	
 		Resources = {