+++ XFaction/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Implemented try/catch/finally exception handling for all timers/events/inputs
 - Some refactoring to old non-C APIs for Classic support (incomplete)
 - Added Evoker/Dracthyr support, awaiting icons for specs
+- Prompts user when it detects a newer version available
 - Added Metrics (X) DT that displays the following total/average values:
   - Total messages received
   - BNet messages received